move client control functions into correct header
