+++ src/client/client_engine.h
@@ -2,11 +2,8 @@
 
 struct ClientConfig;
 
-#include "engine_status.h"
-
 /**
  * @brief Run the client
  * @param config
- * @return EngineStatus The status of the engine on exit
  */
 void runClientEngine(const ClientConfig& config);
